\name{calves}
\docType{data}
\alias{calves}
\title{Delivery of Dutch Calves}
\description{
   This data set gives a cross classification of 7275
   calves born in the late nineties according to type
   of production and type of delivery.
}
\usage{data(calves)}
\format{A data frame containing a contingency table of 7275 observations.}
\source{Holland Genetics.  \url{http://www.hg.nl} }
\references{
   Graffelman, J. (2005) \emph{A guide to scatterplot and biplot calibration}.
}
\keyword{datasets}