\name{goblets}
\docType{data}
\alias{goblets}
\title{Size measurements of archeological goblets}
\description{
   This data set gives 6 different size measurements of 25 goblets
}
\usage{data(goblets)}
\format{A data frame containing 25 observations.}
\source{Manly, 1989}
\references{
   Manly, B. F. J. (1989) \emph{Multivariate statistical methods: a primer}.
   London: Chapman and Hall, London
}
\keyword{datasets}
     