\name{heads}
\docType{data}
\alias{heads}
\title{Dimensions of heads of first and second sons for 25 families}
\description{
   Variables X1 and X2 are the head length and head breadth of the first son and 
   Y1 and Y2 are the same variables for the second son. 
}
\usage{data(heads)}
\format{A data frame containing 25 observations.}
\source{Mardia, 1979, p. 121}
\references{
   Frets, G. P. (1921) \emph{Heredity of head form in man},
   Genetica 3, pp. 193-384.

   Mardia, K. V. and Kent, J. T. and Bibby, J. M. (1979) \emph{Multivariate Analysis}.
   Academic Press London.

   Anderson, T. W. (1984) \emph{An Introduction to Multivariate Statistical Analysis}.
   New York: John Wiley, Second edition.
}
\keyword{datasets}
     