\name{linnerud}
\docType{data}
\alias{linnerud}
\title{Linnerud's exercise and body measurements}
\description{
   The data set consist of 3 exercise variables (Tractions à la barre
   fixe, Flexions, Sauts) and 3 body measurements (Poids, Tour de talle,
   Pouls) of 20 individuals.
}
\usage{data(linnerud)}
\format{A data frame containing 20 observations.}
\source{Tenenhaus, 1998, table 1, page 15}
\references{
   Tenenhaus, M. (1998) \emph{La R\'{e}gression PLS}.
   Paris: Editions Technip.
}
\keyword{datasets}
     