\name{spaindist}
\docType{data}
\alias{spaindist}
\title{Road distances between Spanish cities}
\description{
   Road distances in kilometers between 47 Spanish cities
}
\usage{data(spaindist)}
\format{A data frame containing 47 observations.}
\references{
  Graffelman, J. (2019) Goodness-of-fit filtering in classical metric multidimensional scaling with large
  datasets. <doi: 10.1101/708339>
 }
\keyword{datasets}
     