\name{storks}
\docType{data}
\alias{storks}
\title{Frequencies of nesting storks in Denmark}
\description{
   Danish data from 1953-1977 giving the frequency of nesting storks,
   the human birth rate and the per capita electricity consumption.
}
\usage{data(storks)}
\format{A data frame containing 25 observations.}
\source{Gabriel and Odoroff, Table 1.}
\references{
   Gabriel, K. R. and Odoroff, C. L. (1990) Biplots in biomedical research.
   \emph{Statistics in Medicine} 9(5): pp. 469-485.
}
\keyword{datasets}
     